Create fileCompressor.c
buildMode(): build the mode of operation: store the flags and filenames in struct flags, and check for input error
checkHCZ(): checks if a file ends in .hcz
updatePathname(): add a new directory name to the current pathname, return the new path
updatePathSize(): reallocate memory for the pathsize if needed
err(): write the specified error to STDERR
checkDir(): returns DIRECTORY if the path of the given pathname+filename is a directory
checkFile(): returns 1 if the filename is a regular file. The file must be in the current working directory!! (intended to be used to check the name, not the pathname)
findLast(): return the last directory in the pathname. Must be called on a copy of the pathname, it will change the memory.